# docs/swagger.yml
swagger: '2.0'
info:
  title: Demo API
  description: Happy to code Demo API
  version: 1.0.0
schemes:
  - http
  - https
basePath: /api/v1
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /demo/hello:
    get:
      summary: Hello API
      description: |
        Hello API
      parameters:
        - name: token
          in: header
          type: string
          description: token key
        - name: name
          in: query
          description: Name which to say hello
          required: false
          type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal error
      tags:
        - Hello

  /demo/demo_db:
    get:
      summary: Get data from PostgreSQL API
      description: |
        Get data from PostgreSQL API
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal error
      tags:
        - Hello

  /auth/register:
    post:
      summary: Register new user
      description: |
        Register new user
      parameters:
        - name: body
          in: body
          description: json body data to add new user
          required: true
          schema:
            required:
              - email
              - password
            type: object
            description: User Model
            properties:
              email:
                type: string
                description: email
              password:
                type: string
                description: password
              first_name:
                type: string
                description: first_name
              last_name:
                type: string
                description: last_name
              type:
                type: string
                description: user type
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal error
      tags:
        - Auth
  /auth/login:
    post:
      summary: Login APi
      description: |
        Login API
      parameters:
        - name: body
          in: body
          description: json body data to login
          required: true
          schema:
            required:
              - email
              - password
            type: object
            description: Login Model
            properties:
              email:
                type: string
                description: email
              password:
                type: string
                description: password
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal error
      tags:
        - Auth